Changed layout to support two Webapps: guest page & user page
