+++ components/guest/App.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const App = () => {
+  return <> </>;
+};
