+++ components/user/DashBoard.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export const UserBoard = () => {
+export const DashBoard = () => {
   return (
     <>
       <h5>User Board Page</h5>
